feat: Refactor main bot file to use modular system
